Tidy terminology components
