+++ src/l10n-tools/manager/config/tools/terminology/forms/import.js
@@ -3,33 +3,22 @@ import React from 'react';
 
 import Table from 'web-ext-plugins/widgets/table';
 
+import columns from './columns';
+
 
 export default class TerminologyImportTable extends React.Component {
 
+    get columns () {
+        return columns
+    }
+
     render() {
-        const columns = [
-            {columns: [
-                {Header: "tableHeaderTerm",
-                 id: "term",
-                 accessor: t => t.term
-                },
-                {
-                    Header: "tableHeaderTranslation",
-                    id: "translation",
-                    accessor: t => t.translation
-                },
-                {
-                    Header: "tableHeaderDescription",
-                    id: "description",
-                    accessor: t => t.description
-                },
-            ]}];
         return (
             <div>
               <Table
                  defaultPageSize={5}
                  className="-striped -highlight"
-                 columns={columns}
+                 columns={this.columns}
                  data={this.props.data} />
             </div>);
     }
